reuse getWeeks
