+++ packages/utilities/src/dateTime/components/CalendarDayGrid.ts
@@ -1,6 +1,7 @@
-import { compareDates, compareDatePart, addDays, getDateRangeArray } from '../dateMath/DateMath';
+import { compareDates, compareDatePart, addDays, getDateRangeArray, isInDateRangeArray } from '../dateMath/DateMath';
 import { DayOfWeek, DateRangeType } from '../dateValues/DateValues';
 import { find } from '../../array';
+import { ICalendarStrings } from './Calendar';
 
 export const DAYS_IN_WEEK = 7;
 
@@ -81,6 +82,11 @@ export interface ICalendarDayGridProps {
    * If set the Calendar will not allow selection of dates in this array.
    */
   restrictedDates?: Date[];
+
+  /**
+   * Localized strings to use in the Calendar
+   */
+  strings: ICalendarStrings;
 
   /**
    * The days that are selectable when `dateRangeType` is WorkWeek.
@@ -209,4 +215,98 @@ export function getDateRangeTypeToUse(
   }
 
   return dateRangeType;
+}
+
+/**
+ * Initial parsing of the given props to generate IDayInfo two dimensional array, which contains a representation
+ * of every day in the grid. Convenient for helping with conversions between day refs and Date objects in callbacks.
+ */
+export function getWeeks(props: ICalendarDayGridStateAndProps, onSelectDate: Function): IDayInfo[][] {
+  const {
+    selectedDate,
+    dateRangeType,
+    firstDayOfWeek,
+    today,
+    minDate,
+    maxDate,
+    weeksToShow,
+    workWeekDays,
+    daysToSelectInDayView,
+  } = props;
+
+  const todaysDate = today || new Date();
+
+  const navigatedDate = props.navigatedDate ? props.navigatedDate : todaysDate;
+
+  let date;
+  if (weeksToShow && weeksToShow <= 4) {
+    // if showing less than a full month, just use date == navigatedDate
+    date = new Date(navigatedDate.toString());
+  } else {
+    date = new Date(navigatedDate.getFullYear(), navigatedDate.getMonth(), 1);
+  }
+  const weeks: IDayInfo[][] = [];
+
+  // Cycle the date backwards to get to the first day of the week.
+  while (date.getDay() !== firstDayOfWeek) {
+    date.setDate(date.getDate() - 1);
+  }
+
+  // add the transition week as last week of previous range
+  date = addDays(date, -DAYS_IN_WEEK);
+
+  // a flag to indicate whether all days of the week are outside the month
+  let isAllDaysOfWeekOutOfMonth = false;
+
+  // in work week view if the days aren't contiguous we use week view instead
+  const selectedDateRangeType = getDateRangeTypeToUse(dateRangeType, workWeekDays);
+
+  let selectedDates = getDateRangeArray(
+    selectedDate,
+    selectedDateRangeType,
+    firstDayOfWeek,
+    workWeekDays,
+    daysToSelectInDayView,
+  );
+  selectedDates = getBoundedDateRange(selectedDates, minDate, maxDate);
+
+  let shouldGetWeeks = true;
+
+  for (let weekIndex = 0; shouldGetWeeks; weekIndex++) {
+    const week: IDayInfo[] = [];
+
+    isAllDaysOfWeekOutOfMonth = true;
+
+    for (let dayIndex = 0; dayIndex < DAYS_IN_WEEK; dayIndex++) {
+      const originalDate = new Date(date.toString());
+      const dayInfo: IDayInfo = {
+        key: date.toString(),
+        date: date.getDate().toString(),
+        originalDate,
+        isInMonth: date.getMonth() === navigatedDate.getMonth(),
+        isToday: compareDates(todaysDate, date),
+        isSelected: isInDateRangeArray(date, selectedDates),
+        onSelected: () => {
+          onSelectDate(originalDate);
+        },
+        isInBounds: !getIsRestrictedDate(props, date),
+      };
+
+      week.push(dayInfo);
+
+      if (dayInfo.isInMonth) {
+        isAllDaysOfWeekOutOfMonth = false;
+      }
+
+      date.setDate(date.getDate() + 1);
+    }
+
+    // We append the condition of the loop depending upon the showSixWeeksByDefault prop.
+    shouldGetWeeks = weeksToShow ? weekIndex < weeksToShow + 1 : !isAllDaysOfWeekOutOfMonth || weekIndex === 0;
+
+    // we don't check shouldGetWeeks before pushing because we want to add one extra week for transition state
+    weeks.push(week);
+  }
+
+  return weeks;
 }
